Create README.md
"Resturan Genie" is an app that will help its users with their different needs
The customer will be able to view his order from the moment he order it untill it is ready.
The waiter will be able to balance the orders and respond efficiently. The shief will also get a better sorted list of orders, and of course, the resturant manger will will be able to suprvise his resturant all the time, giving him the ability to quickly solve any problems that may occur.